refactor validation code
