+++ tools/dspl2/dspl2/rdfutil.py
