+++ tools/dspl2/dspl2/templates/viewer.css
