--- tools/dspl2/dspl2/validator/__init__.py
@@ -1,5 +1,0 @@
-# Copyright 2018 Google LLC
-#
-# Use of this source code is governed by a BSD-style
-# license that can be found in the LICENSE file or at
-# https://developers.google.com/open-source/licenses/bsd
