+++ tools/dspl2/scripts/validator.py
@@ -10,11 +10,11 @@ import jinja2
 from pathlib import Path
 import sys
 
-import dspl2.validator
-from dspl2.validator.expander import ExpandStatisticalDataset
-from dspl2.validator.filegetter import *
-from dspl2.validator.jsonutil import JsonToKwArgsDict
-from dspl2.validator.rdfutil import NormalizeJsonLd
+import dspl2
+from dspl2.expander import ExpandStatisticalDataset
+from dspl2.filegetter import *
+from dspl2.jsonutil import JsonToKwArgsDict
+from dspl2.rdfutil import NormalizeJsonLd
 
 
 FLAGS = flags.FLAGS
@@ -23,7 +23,7 @@ flags.DEFINE_boolean('normalize', False,
 
 
 def RenderLocalDspl2(path, normalize):
-  template_dir = Path(dspl2.validator.__file__).parent / 'templates'
+  template_dir = Path(dspl2.__file__).parent / 'templates'
   env = jinja2.Environment(loader=jinja2.FileSystemLoader(
       template_dir.as_posix()))
   try:
